Move SmartSlider to themed directory
