+++ src/components/themed/SmartSlider.js
@@ -6,10 +6,10 @@ import { PanGestureHandler } from 'react-native-gesture-handler'
 import Animated, { Easing, runOnJS, useAnimatedGestureHandler, useAnimatedStyle, useSharedValue, withTiming } from 'react-native-reanimated'
 import Entypo from 'react-native-vector-icons/Entypo'
 
-import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../../../../components/services/ThemeContext.js'
-import { EdgeText } from '../../../../components/themed/EdgeText'
-import s from '../../../../locales/strings.js'
-import { useCallback, useState } from '../../../../types/reactHooks.js'
+import s from '../../locales/strings.js'
+import { useCallback, useState } from '../../types/reactHooks.js'
+import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
+import { EdgeText } from './EdgeText'
 
 const COMPLETE_POINT: number = 3
